use a webworker to calculate vertices position
